Improve help text and change some other text
I added some more information to the help text at the beginning of
the file. Now it is in full accordance with the README.md file.

I also changed text at a different place because I felt that the
plural form of the word "statement" could be appropriate as well.