Add new option: show first n chars of DB name
A new option has been added to the configuration block. The user
can now set a limit to the length of database names to be shown.

This new option further enhances the usability of MDBM by providing
a nice way to align database names in the list of selected databases.